refactor(ssh): extract duplicate SSH resolution to common/ssh_bridge.rs
Removed resolve_ssh_in_config from agent_adapters.rs and resolve_config_via_ssh from capabilities/commands.rs. Both now call the shared resolve_ssh_in_place(app, config). extract_remote_target moved to ssh_bridge.rs with its tests.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
